Move org.wso2.carbon.ui.filters.cache classes to org.wso2.carbon.tomcat.ext.filter
