+++ core/org.wso2.carbon.tomcat.ext/src/main/java/org/wso2/carbon/tomcat/ext/filter/AbstractCachePreventionFilter.java
@@ -1,22 +1,22 @@
 /*
-*  Copyright (c) 2016, WSO2 Inc. (http://www.wso2.org) All Rights Reserved.
-*
-*  WSO2 Inc. licenses this file to you under the Apache License,
-*  Version 2.0 (the "License"); you may not use this file except
-*  in compliance with the License.
-*  You may obtain a copy of the License at
-*
-*    http://www.apache.org/licenses/LICENSE-2.0
-*
-* Unless required by applicable law or agreed to in writing,
-* software distributed under the License is distributed on an
-* "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
-* KIND, either express or implied.  See the License for the
-* specific language governing permissions and limitations
-* under the License.
-*/
+ * Copyright (c) 2023, WSO2 LLC. (https://www.wso2.com) All Rights Reserved.
+ *
+ * WSO2 LLC. licenses this file to you under the Apache License,
+ * Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
 
-package org.wso2.carbon.ui.filters.cache;
+package org.wso2.carbon.tomcat.ext.filter; 
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -47,11 +47,6 @@ import org.apache.commons.logging.LogFactory;
  *
  * @since 4.2.0
  */
-
-/**
- * @deprecated This class is moved to org.wso2.carbon.tomcat.ext
- */
-@Deprecated
 public abstract class AbstractCachePreventionFilter implements Filter {
 
     private static Log log = LogFactory.getLog(AbstractCachePreventionFilter.class);
