+++ distribution/kernel/carbon-home/repository/conf/tomcat/carbon/WEB-INF/web.xml
@@ -87,7 +87,7 @@
 
     <filter>
         <filter-name>URLBasedCachePreventionFilter</filter-name>
-        <filter-class>org.wso2.carbon.ui.filters.cache.URLBasedCachePreventionFilter</filter-class>
+        <filter-class>org.wso2.carbon.tomcat.ext.filter.URLBasedCachePreventionFilter</filter-class>
     </filter>
 
     <filter-mapping>
